### 2025-09-26 d24ce8c
Add De Morgan's law for Pred
