### 2025-09-26 0a974c5
Add more of De Morgan's laws
