### 2025-09-26 1bcf673
Savagely cut for length. Not quite there yet but a lot closer
