+++ paper/paper.md
@@ -92,36 +92,28 @@ bibliography: paper.bib
 # Summary
 
 Agda [@agda2024manual] is a dependently-typed functional
-language that serves both as a traditional programming language
-and as an interactive theorem prover (ITP).
-In other words, its type system is expressive enough to formulate
-complex formal requirements on programs as types, and its compiler allows users to interactively build terms and check that they satisfy these requirements.
-Through the Curry-Howard lens [@DBLP:journals/cacm/Wadler15],
-these types and terms can be seen respectively as theorem
-statements and proofs.
+language that serves as both a programming language
+and an interactive theorem prover (ITP).
+More precisely, one can formulate requirements on programs as types, and building programs satisfying these requirements is performed in a fundamentally interactive manner.
+The Curry-Howard lens [@DBLP:journals/cacm/Wadler15], types and terms can be seen as theorems and their proofs.
 
-This paper presents the Agda standard library [@agda-stdlib-v2.0], hereafter referred to as `agda-stdlib`, which provides definitions intended to be helpful for users to develop Agda programs and proofs.
-Unlike the standard libraries of traditional programming languages, `agda-stdlib` provides not only standard utilities and data structures, but also a range of basic discrete mathematics useful for proving the correctness of programs.
+We present the Agda standard library [@agda-stdlib-v2.0], hereafter referred to as `agda-stdlib`, which provides definitions intended to be generally helpful in developments of programs and proofs.
+Unlike standard libraries of traditional programming languages, `agda-stdlib` provides not only standard utilities and data structures, but also a range of basic discrete mathematics useful for proving the correctness of programs.
 
 # Statement of need
 
-Most programming languages include a "standard" library offering a basic set of algorithms, data structures, and operating system procedures.
-However, there are two reasons why a standard library is particularly significant for Agda compared to traditional programming languages.
+There are two reasons why a standard library for Agda is different than for traditional languages.
 
-First, like other theorem provers, the Agda language provides only a small set of primitives from which programs can be constructed.
-As a result, many concepts traditionally considered part of a language must be defined within the program itself.
-This approach reduces compiler complexity and enhances its reliability, and also demonstrates the strength of the core Agda language as it can push these concepts out to the library.
-For example, in a fresh Agda environment, there is no predefined notion of an integer, let alone more complex data structures such as vectors or maps.
-This lack of basic data types increases the need for a standard library when compared to more mainstream languages.
+First, Agda's language is small but sufficient to define many concepts usually considered part of a language (such
+as the notion of an integer).
+This reduces compiler complexity. But it increases the need for a standard library.
 
-Second, Agda users often seek to prove that programs constructed using data types in the standard library are "correct."
-Constructing the proof that a function obeys a specification (e.g. that a sorting function outputs a permutation of the original list in non-decreasing order) often requires more effort, both in terms of lines of code and in developer time, than writing the original operation.
-By providing proofs of correctness for operations it defines, the standard library saves users significant time during proof development.
+Second, as correctness is an important goal of most Agda users, it is important that functions provided by `agda-stdlib` come with correctness guarantees. Such proofs often require significant effort which should not be downloaded to users.
 
 # Impact
 
 A wide range of projects make use of `agda-stdlib`.
-In the list below we present a selection of such projects:
+Here is a selection:
 
 - Programming Language Foundations in Agda [@plfa22.08]
 
@@ -135,91 +127,76 @@ In the list below we present a selection of such projects:
 
 - Verification of routing protocols [@daggitt2023routing]
 
[…]
