### 2025-09-30 40beb67
More distributivity properties for nats (#2833)
* More distributivity properties for nats

* Simplify proof using existing lemma

* Use lemma instead of refl

In principle this lets us translate more easily to an arbitrary semiring
