### 2025-10-01 ea2a7ce
[ add ] flipped version of `Relation.Nullary.Negation.Core.contradiction` (#2797)
* add: flipped version of `contradiction`

* add: comment to `CHANGELOG`
