+++ paper/paper.md
@@ -20,7 +20,7 @@ authors:
     orcid: 0000-0003-0420-4492
     affiliation: 4
   - name: van Doorn, Nathan
-    orcid: 0009-0009-0598-3663
+    orcid: 0009-0009-0598-3636
     affiliation: 5
   - name: Wood, James
     orcid: 0000-0002-8080-3350
