### 2025-10-04 bc9aac7
Added missing DOI
