+++ paper/paper.bib
@@ -192,7 +192,8 @@ location = {Bertinoro, Italy}
   title={Isabelle: A generic theorem prover},
   author={Paulson, Lawrence C},
   year={1994},
-  publisher={Springer}
+  publisher={Springer},
+  doi={10.1007/BFb0030541}
 }
 
 @misc{agda2024manual,
