+++ paper/paper.bib
@@ -4,6 +4,7 @@
     year   = {2022},
     month  = aug,
     url    = {https://plfa.inf.ed.ac.uk/22.08/},
+    doi    = {10.1016/j.scico.2020.102440}
 }
 
 @mastersthesis{ivardeBruin2023,
