### 2025-10-04 0435c49
More shortening
