+++ paper/paper.md
@@ -91,60 +91,53 @@ bibliography: paper.bib
 
 # Summary
 
-Agda [@agda2024manual] is a dependently-typed functional
-language that serves as both a programming language
-and an interactive theorem prover (ITP).
-More precisely, one can formulate requirements on programs as types, and building programs satisfying these requirements is performed in a fundamentally interactive manner.
-The Curry-Howard lens [@DBLP:journals/cacm/Wadler15], types and terms can be seen as theorems and their proofs.
+Agda [@agda2024manual] is a dependently-typed functional language that serves as both a programming language and an interactive theorem prover (ITP).
+In Agda, one can formulate requirements on programs as types, and build programs satisfying these requirements interactively.
+The Curry-Howard correspondance [@DBLP:journals/cacm/Wadler15] allows types and programs to be seen as theorems and proofs.
 
-We present the Agda standard library [@agda-stdlib-v2.0], hereafter referred to as `agda-stdlib`, which provides definitions intended to be generally helpful in developments of programs and proofs.
-Unlike standard libraries of traditional programming languages, `agda-stdlib` provides not only standard utilities and data structures, but also a range of basic discrete mathematics useful for proving the correctness of programs.
+We present the Agda standard library [@agda-stdlib-v2.0] (`agda-stdlib`), which provides definitions helpful in the development of programs and proofs.
+Unlike standard libraries of traditional programming languages, `agda-stdlib` provides not only standard utilities and data structures, but also basic discrete mathematics useful for proving the correctness of programs.
 
 # Statement of need
 
-There are two reasons why a standard library for Agda is different than for traditional languages.
+Aside from the normal justifications for the existence of standard libraries, there are two reasons why `agda-stdlib` is uniquely needed.
 
-First, Agda's language is small but sufficient to define many concepts usually considered part of a language (such
-as the notion of an integer).
-This reduces compiler complexity. But it increases the need for a standard library.
+First, Agda as a language is small and does not include concepts usually considered part of a language (such as integers and strings).
+This reduces compiler complexity, but means that `agda-stdlib` must define such concepts itself.
 
-Second, as correctness is an important goal of most Agda users, it is important that functions provided by `agda-stdlib` come with correctness guarantees. Such proofs often require significant effort which should not be downloaded to users.
+Second, as program correctness is important for most Agda users, the functions provided by `agda-stdlib` come with correctness guarantees. Such proofs often require significant effort which should not be offloaded to users.
 
 # Impact
 
-A wide range of projects make use of `agda-stdlib`.
-Here is a selection:
+A wide range of projects make use of `agda-stdlib`, including:
 
 - Programming Language Foundations in Agda [@plfa22.08]
 
 - Formalisation of category theory [@hu2021categories]
 
-- Intrinsically typed interpreters for imperative languages [@bach2017intrinsically] and formalisation of type-level computation and subtyping in Scala [@stucki2021theory]
+- Formalisation of aspects of Scala's type system [@stucki2021theory]
 
-- Formally verified calculus for the synchronous, reactive programming language Esterel [@florence2019esterel]
+- Verification of a calculus for the synchronous, reactive programming language Esterel [@florence2019esterel]
 
 - Verification of hardware circuit design [@pizani2018pi]
 
 - Verification of routing protocols [@daggitt2023routing]
 
-`agda-stdlib` has also been used as a test bed for the design of co-inductive data types in Agda itself, resulting having in three different notions.
-
-The development of `agda-stdlib` has also had a synergistic relationship with that of Agda itself, prompting new language features.
-Firstly, Agda supports a wide range of possibly incompatible language extensions.
[…]
