### 2025-10-04 9cb6502
More shortening
