+++ paper/paper.md
@@ -117,62 +117,54 @@ A wide range of projects make use of `agda-stdlib`, including:
 
 - Formalisation of aspects of Scala's type system [@stucki2021theory]
 
-- Verification of a calculus for the synchronous, reactive programming language Esterel [@florence2019esterel]
+- Verification of a calculus for the synchronous, reactive Esterel language [@florence2019esterel]
 
 - Verification of hardware circuit design [@pizani2018pi]
 
 - Verification of routing protocols [@daggitt2023routing]
 
 The development of `agda-stdlib` has also had a synergistic relationship with that of Agda itself, helping to both test and motivate new language features.
-Firstly, Agda supports a wide range of possibly incompatible language extensions, with examples including `--cubical` (changing the underlying type theory to cubical type theory [@DBLP:journals/jfp/VezzosiMA21]),
-`--with-K` (adding support for Streicher's axiom K [@streicher1993investigations]),
-or `--safe` (an ITP-oriented option restricting sources of potential unsoundness).
-In order for `agda-stdlib` to be compatible with as many different combinations of extensions as possible, it is designed modularly with units
-requesting the minimal expressive power needed.
+For example, Agda supports a wide range of incompatible language extensions and `agda-stdlib` is designed modularly in order to be compatible with many different combinations of extensions.
+Modules then request the minimal expressive power needed.
 To facilitate this, Agda's extensions were categorised as "infective", "coinfective" or "neither", where
 "infective" options impact all the import*ing* modules and "coinfective" options impact the import*ed* modules.
-Secondly, another feature motivated by of `agda-stdlib` is attaching custom messages to definitions, displayed on use.
-This enabled deprecation warnings, easing the evolution of user code.
-The library has also been used as a test bed for different approaches to defining co-inductive data types in Agda itself.
+The library has also acted as a test bed for different approaches to defining co-inductive data types in Agda.
 
 # Design Challenges
 
 Organising libraries of discrete mathematics and algebra coherently is notoriously difficult
 [@carette2020leveraging] [@cohen2020hierarchy].
-There is a tension between organising the material to be as general as possible and providing direct and intuitive definitions.
-Mathematical objects often permit multiple representations with varying application-specific usability profiles -- but this leads to redundancy/duplication.
-Some theorem provers ([@coq2024manual], [@paulson1994isabelle]) choose instead to have a rich ecosystem of external libraries, reducing the pressure to have canonical definitions for common concepts at the risk of incompatibilities between libraries.
+There is a tension between being as general as possible and providing direct and intuitive definitions.
+Mathematical objects often permit multiple representations with different advantages -- but this leads to redundancy/duplication.
+Some ITPs ([@coq2024manual], [@paulson1994isabelle]) have a rich ecosystem of external libraries, avoiding canonical definitions at the risk of incompatibilities between libraries.
 We have chosen, like Lean's `mathlib` [@van2020maintaining], to provide a repository of canonical definitions.
 
-`agda-stdlib` has embraced the "intrinsic style" of dependent types, in which correctness invariants are part of the data structures themselves.
-For instance, the definition of rational numbers carries a proof showing that the numerator and denominator have no common factors.
-The intrinsic style also means returning proofs from decision procedures rather than booleans.
+`agda-stdlib` has embraced the "intrinsic style" of dependent types, where data structures themselves contain correctness invariants.
+For instance, rational numbers include a proof that the numerator and denominator have no common factors and decision procedures return proofs rather than booleans.
 We believe that `agda-stdlib` has been one of the first ITP standard libraries to tackle this challenge.
 
-The library has also embraced Agda's parametrised modules to enable polymorphism [@ivardeBruin2023].
[…]
