### 2025-10-04 196af9f
More shortening
