### 2025-10-04 9ba8ff5
More shortening
