### 2025-10-04 3e04ff4
refactor: simplify
