### 2025-10-05 669189d
[ add ] `BooleanRing` plus `Properties` (#2763)
* add: `BooleanRing` plus `Properties`

* Update CHANGELOG.md

Removed spurious entry

* refactor: more general API

* fix: added note to `CHANGELOG`

* refactor: make binomial expansion an explicit step

* rename: `*-isIdempotentMonoid`

* add: `*-idempotentMonoid`

* refactor: use `Cancellative` properties only

* add: stub properties for `CommutativeRing`

* refactor: `BooleanRing` properties in terms of `CommutativeRing` and `BooleanSemiring`

* refactor: put everything together

* add: stub properties for `CommutativeRing`

* refactor: adjus
[…]
