+++ CHANGELOG.md
@@ -47,6 +47,14 @@ Deprecated names
 New modules
 -----------
 
+* `Algebra.Properties.BooleanRing`.
+
+* `Algebra.Properties.BooleanSemiring`.
+
+* `Algebra.Properties.CommutativeRing`.
+
+* `Algebra.Properties.Semiring`.
+
 * `Data.List.Relation.Binary.Permutation.Algorithmic{.Properties}` for the Choudhury and Fiore definition of permutation, and its equivalence with `Declarative` below.
 
 * `Data.List.Relation.Binary.Permutation.Declarative{.Properties}` for the least congruence on `List` making `_++_` commutative, and its equivalence with the `Setoid` definition.
@@ -54,10 +62,41 @@ New modules
 Additions to existing modules
 -----------------------------
 
+* In `Algebra.Bundles`:
+  ```agda
+  record BooleanSemiring _ _ : Set _
+  record BooleanRing _ _     : Set _
+  ```
+
+* In `Algebra.Consequences.Propositional`:
+  ```agda
+  binomial-expansion : Associative _∙_ → _◦_ DistributesOver _∙_ →
+    ∀ w x y z → ((w ∙ x) ◦ (y ∙ z)) ≡ ((((w ◦ y) ∙ (w ◦ z)) ∙ (x ◦ y)) ∙ (x ◦ z))
+  ```
+
+* In `Algebra.Consequences.Setoid`:
+  ```agda
+  binomial-expansion : Congruent₂ _∙_  → Associative _∙_ → _◦_ DistributesOver _∙_ →
+    ∀ w x y z → ((w ∙ x) ◦ (y ∙ z)) ≈ ((((w ◦ y) ∙ (w ◦ z)) ∙ (x ◦ y)) ∙ (x ◦ z))
+  ```
+
+* In `Algebra.Lattice.Properties.BooleanAlgebra.XorRing`:
+  ```agda
+  ⊕-∧-isBooleanRing : IsBooleanRing _⊕_ _∧_ id ⊥ ⊤
+  ⊕-∧-booleanRing   : BooleanRing _ _
+  ```
+
 * In `Algebra.Properties.RingWithoutOne`:
   ```agda
   [-x][-y]≈xy : ∀ x y → - x * - y ≈ x * y
   ```
+
+* In `Algebra.Structures`:
+  ```agda
+  record IsBooleanSemiring + * 0# 1# : Set _
+  record IsBooleanRing + * - 0# 1# : Set _
+  ```
+  NB. the latter is based on `IsCommutativeRing`, with the former on `IsSemiring`.
 
 * In `Data.Fin.Permutation.Components`:
   ```agda
