### 2025-10-05 8d9f7de
[bug] replace wrong argument to projection in `Algebra.Morphism.Construct.DirectProduct` (#2822)
* [bug] fix "wrong" argument to projection

* add: note on bug fix to `CHANGELOG`

---------

Co-authored-by: jamesmckinna <j.mckinna@hw.ac.uk>
