### 2025-10-07 #9071 @JacquesCarette
(I'm not @MatthewDaggitt but I am one of the authors of the paper. In an effort to speed things along, I'll suggest some potential reviewers from that list.)

In order of most-qualified to least-qualified (but still roughly able to do it.)
- Dominic Orchard.
- Louise A. Dennis.
- Kiran Gopinathan
- Vindush Singhal
- Luca Ferranti

As far as people not on the list but who definitely could review, on par with the first reviewer above:
- Joey Eremondi (JoeyEremondi)
- Cass Alexandru (csandru)
- Philip Wadler (wadler)
