### 2025-10-07 a53bce8
Update CI to Agda-2.8.0 and GHC-9.12.2 (shipped with ubuntu-24.04)
