### 2025-10-07 c233a37
Update CI to Agda-2.8.0 and GHC-9.12.2 (#2835)
* Update stack.yaml files to latest minor versions

* Update CI to Agda-2.8.0 and GHC-9.12.2 (shipped with ubuntu-24.04)
