+++ stack-9.6.2.yaml
@@ -1,6 +1,0 @@
-resolver: nightly-2023-08-27
-compiler: ghc-9.6.2
-compiler-check: match-exact
-
-packages:
-- '.'
