### 2025-10-07 #9071 @danielskatz
👋  @vidsinghal & @lucaferranti, would one or both of you be willing to review this submission for JOSS? We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/review_criteria.html
