### 2025-10-07 #9071 @lucaferranti
I would love too, but unfortunately I'm pretty busy atm, and frankly I don't think my agda proficiency is enough for high quality review.

Nevertheless, this is super cool and exciting to see, I wish all the best to the authors!!
