### 2025-10-08 #9071 @MatthewDaggitt
> @MatthewDaggitt - just as a minor point, when you have multiple references, you should use [@ref1; @ref2] in the .md file.

Thanks, have fixed! Should I call `@editorialbot generate pdf` every time I make a tweak?
