### 2025-10-09 0576db4
[ new ] Effect.Monad.Random (#2372)
* [ new ] Effect.Monad.Random

Still TODO: instances for transformers

* [ more ] add lifting along transformer

* [ test ] for the new mtl-style constraint

* [ cleanup ] this is not meant to be uploaded

* [ fix ] declare effect.monad.random as unsafe

* [ changelog ] listing new modules too

* fix: `CHANGELOG`

---------

Co-authored-by: Jacques Carette <carette@mcmaster.ca>
Co-authored-by: jamesmckinna <j.mckinna@hw.ac.uk>
Co-authored-by: jamesmckinna <31931406+jamesmckinna@users.noreply.github.com>
