+++ CHANGELOG.md
@@ -61,6 +61,8 @@ New modules
 
 * `Data.List.Relation.Binary.Permutation.Declarative{.Properties}` for the least congruence on `List` making `_++_` commutative, and its equivalence with the `Setoid` definition.
 
+* `Effect.Monad.Random` and `Effect.Monad.Random.Instances` for an mtl-style randomness monad constraint.
+
 Additions to existing modules
 -----------------------------
 
@@ -155,4 +157,10 @@ Additions to existing modules
   ```agda
   ¬¬-η           : A → ¬ ¬ A
   contradiction′ : ¬ A → A → Whatever
+  ```
+
+* In `System.Random`:
+  ```agda
+  randomIO : IO Bool
+  randomRIO : RandomRIO {A = Bool} _≤_
   ```
