### 2025-10-10 2ce5eb3
[Add] truncate properties to Data.Vec.Properties (#2795)
* adding truncate's properties

* update changelog

* cleaning whitespaces

* Recleaning

* corrections after reviews

* correction of changelog

* changes after review
