### 2025-10-13 #9241 @editorialbot
[REVIEW]: The Agda standard library: version 2.0
**Submitting author:** @MatthewDaggitt (<a href="http://orcid.org/0000-0002-2552-3671">Matthew Daggitt</a>)
**Repository:** https://github.com/agda/agda-stdlib
**Branch with paper.md** (empty if default branch): v2.0-joss-submission
**Version:** v2.0-joss-paper
