+++ CHANGELOG.md
@@ -180,6 +180,11 @@ Additions to existing modules
   contradiction′ : ¬ A → A → Whatever
   ```
 
+* In `Relation.Unary`
+  ```agda
+  _⊣_ : (A → B) → Pred A ℓ → Pred B _
+  ```
+
 * In `System.Random`:
   ```agda
   randomIO : IO Bool
