+++ CHANGELOG.md
@@ -13,6 +13,8 @@ Bug-fixes
 
 * Fix a typo in `Algebra.Morphism.Construct.DirectProduct`.
 
+* Fix a typo in `Function.Construct.Constant`.
+
 Non-backwards compatible changes
 --------------------------------
 
