### 2025-10-25 450f64f
refactor: drastic streamlining
