### 2025-10-26 7d269ca
[ refactor ] reorganise `Relation.Nullary.Negation.Core` to make clean separation of properties (#2805)
* refactor: reorganise to make clean separation of properties

* refactor: remove `note`

* refactor: remove import of `flip`
