### 2025-10-26 0c71ce1
deprecate: names for deMorgan lemmas (#2837)
