+++ CHANGELOG.md
@@ -53,6 +53,15 @@ Deprecated names
   interchange  ↦   medial
   ```
 
+* In `Relation.Nullary.Decidable.Core`:
+  ```agda
+  ⊤-dec     ↦   ⊤?
+  ⊥-dec     ↦   ⊥?
+  _×-dec_  ↦   _×?_
+  _⊎-dec_  ↦   _⊎?_
+  _→-dec_  ↦   _→?_
+  ```
+
 * In `Relation.Nullary.Negation`:
   ```agda
   ∃⟶¬∀¬  ↦   ∃⇒¬∀¬
