### 2025-10-26 #9241 @MatthewDaggitt
Thanks @dorchard!

> I am generally happy with everything here apart from one point: it's not clear how to build and run the automated tests. There is a makefile but I couldn't get it to work. Perhaps a README could be included for the test suite?

This is fair! I've opened a [PR](https://github.com/agda/agda-stdlib/pull/2844) with a README in the `tests/` folder.
@danielskatz obviously we can't retroactively add it to the v2.0 release, so I hope it's okay? Would you like us to backport it to the JOSS paper branch as well?

> @MatthewDaggitt Can I also just clarify that this intended to be version 2.0 and not 2.3?

Yes, it is indeed 2.0. We planned to submit much earlier, but discussions with a potential co-author took longer than we had anticipated.

> I thought it might also be nice to use the opportunity in the summary to contextualise a little bit more (just a few more sentences) about the role of interactive theorem provers in mathematics, software engineering, and computer science, for a broader audience.

We had this in, but we had to drastically cut the paper to get in under the word limit for JOSS. We can add these sentences back in but we will breach the word limit again. @danielskatz thoughts?

> Minor writing comments
>
>    l.42 "now categories extensions" -> categorises?
>    l.59 "This allows users specify " -> allows uses to specify
>    l.83 "contributations" spelling!

Fixed all of these.

> Is there missing information about Hu's funding source (l.85)?

I thought it was a little odd too, but that is all he gave us when we was asked. I can follow up again with him.
