+++ stack-9.10.3.yaml
@@ -1,5 +1,5 @@
-snapshot: lts-24.11
-compiler: ghc-9.10.2
+snapshot: lts-24.17
+compiler: ghc-9.10.3
 compiler-check: match-exact
 packages:
 - .
