### 2025-10-27 d84689f
[ add ] clean version of `Data.Fin.Properties.searchMinimalCounterexample` (#2801)
* [ add ] clean version of `Data.Fin.Properties.searchMinimalCounterexample`

* use: `Π[_]` syntax

* refactor: simplify

* refactor: drastic streamlining

* refactor: remove redundant `import`

* refactor: inlined `lemma` in `¬∀⇒∃¬`
