### 2025-10-28 dd64a3a
[ fix ] `Relation.Nullary.Decidable.Core` names for combinators (#2843)
* refactor: fixes #2842

* refactor: knock-ons

* refactor: more knock-ons

* refactor: resolve merge conflict following #2793
