### 2025-10-28 #9241 @benmandrew
@MatthewDaggitt thanks for the documentation. When I try to build the library I get:
```bash
[agda-stdlib]$ agda -Werror +RTS -M5G -H3.5G -A128M -RTS -i. -isrc -idoc -WnoUserWarning Everything.agda
Checking Everything (/home/y19056ba/projects/agda-stdlib/Everything.agda).
 Checking Algebra.Operations.Semiring (/home/y19056ba/projects/agda-stdlib/src/Algebra/Operations/Semiring.agda).
/home/y19056ba/projects/agda-stdlib/src/Algebra/Operations/Semiring.agda:18,1-31,32
Failed to find source of module
Algebra.Properties.Semiring.Exponentiation in any of the following
locations:
  /home/y19056ba/projects/agda-stdlib/src/Algebra/Properties/Semiring/Exponentiation.agda
  /home/y19056ba/projects/agda-stdlib/src/Algebra/Properties/Semiring/Exponentiation.lagda
  /home/y19056ba/projects/agda-stdlib/doc/Algebra/Properties/Semiring/Exponentiation.agda
  /home/y19056ba/projects/agda-stdlib/doc/Algebra/Properties/Semiring/Exponentiation.lagda
  /home/y19056ba/projects/agda-stdlib/Algebra/Properties/Semiring/Exponentiation.agda
  /home/y19056ba/projects/agda-stdlib/Algebra/Properties/Semiring/Exponentiation.lagda
  /home/y19056ba/.cabal/store/ghc-9.6.7/Agda-2.6.4.1-b97a9fd40df3f7b9d805525bef1422fd1a4190f07c8094c2091de4b5b5315e1d/share/lib/prim/Algebra/Properties/Semiring/Exponentiation.agda
  /home/y19056ba/.cabal/store/ghc-9.6.7/Agda-2.6.4.1-b97a9fd40df3f7b9d805525bef1422fd1a4190f07c8094c2091de4b5b5315e1d/share/lib/prim/Algebra/Properties/Semiring/Exponentiation.lagda
when scope checking the declaration
  import Algebra.Properties.Semiring.Exponentiation as .#Algebra.Properties.Semiring.Exponentiation-11535128805283862462
```

(I also have to rename `doc/Everything.agda` otherwise it considers the module name ambiguous.)

For the golden tests, installing `clock` fixed the first issue, but the second still persists --- is there another library I need to install?
```bash
[agda-stdlib]$ make testsuite INTERACTIVE='' AGDA_EXEC='~/.cabal/bin/agda'
make -C tests test AGDA="~/.cabal/bin/agda -Werror +RTS -M4.0G -H3.5G -A128M -RTS" AGDA_EXEC="~/.cabal/bin/agda" only=
make[1]: Entering directory '/home/y19056ba/projects/agda-stdlib/tests'
rm -f _build/runtests
rm -f _build/MAlonzo/Code/Qruntests*
~/.cabal/bin/agda -Werror +RTS -M4.0G -H3.5G -A128M -RTS --compile-dir=_build/ -c runtests.agda
Compiling runtests in /home/y19056ba/projects/agda-stdlib/tests/_build/2.6.4.1/agda/runtests.agdai to _build/MAlonzo/Code/Qruntests.hs
Calling: ghc -O -o _build/runtests -Werror -i_build/ -main-is MAlonzo.Code.Qruntests _build/MAlonzo/Code/Qruntests.hs --make -fwarn-incomplete-patterns
[ 58 of 178] Compiling MAlonzo.Code.System.Directory.Primitive ( _build/MAlonzo/Code/System/Directory/Primitive.hs, _build/MAlonzo/Code/System/Directory/Primitive.o )
Compilation error:

_build/MAlonzo/Code/System/Directory/Primitive.hs:46:5: error: [GHC-62161] [-Wincomplete-patterns, Werror=incomplete-patterns]
    Pattern match(es) are non-exhaustive
    In a case alternative:
        Patterns of type ‘XdgDirectory’ not matched: XdgState
   |
46 |   = case x of
   |     ^^^^^^^^^...

make[1]: *** [Makefile:6: runtests] Error 42
make[1]: Leaving directory '/home/y19056ba/projects/agda-stdlib/tests'
make: *** [GNUmakefile:15: testsuite] Error 2
```
