### 2025-10-31 7a81958
Updated Jason Hu's funding contributions
