### 2025-10-31 #9241 @MatthewDaggitt
Okay so I have:
 - updated Jason Hu's funding contribution statement with more details.
 - cherry-picked commit https://github.com/agda/agda-stdlib/commit/6e23886cffba2cadea20a59af5ce513584866f22 across to the `v2.0-joss-submission` branch (thanks @shhyou for that amazing spot!). That should fix @benmandrew's first error.

The remaining issue is their second error about `XdgDirectory`. To be honest this looks like something going wrong with Agda versions. @benmandrew what version of Agda are you running? `v2.0` of the standard library is compatible with `v2.6.4.X` of Agda.
