### 2025-10-31 969f670
Add normal subgroups and quotient groups
