### 2025-10-31 d02a8cf
add: adjoint properties for #2840
