### 2025-10-31 99c90ae
Bimodule rather than module
