### 2025-10-31 7cb6f93
Add ideals and quotient rings
