+++ HACKING.md
@@ -4,7 +4,7 @@ Contributing to the library
 Thank you for your interest in contributing to the Agda standard library.
 Hopefully this guide should make it easy to do so! Feel free to ask any
 questions on the Agda mailing list. Before you start please read the
-[style-guide](https://github.com/agda/agda-stdlib/blob/master/notes/style-guide.md).
+[style-guide](https://github.com/agda/agda-stdlib/blob/master/doc/style-guide.md).
 
 What is an acceptable contribution?
 ===================================
@@ -124,7 +124,7 @@ git push USER -u new_feature
 
 You can then proceed to make your changes. Please follow existing
 conventions in the library, see
-[style-guide](https://github.com/agda/agda-stdlib/blob/master/notes/style-guide.md).
+[style-guide](https://github.com/agda/agda-stdlib/blob/master/doc/style-guide.md).
 for details. Document your changes in `agda-stdlib-fork/CHANGELOG.md`.
 
 If you are creating new modules, please make sure you are having a
@@ -244,23 +244,14 @@ you are never committing anything with a whitespace violation:
 Type-checking the README directory
 ----------------------------------
 
-* By default the README directory is not exported in the
-  `standard-library.agda-lib` file in order to avoid clashing with other people's
-  README files. This means that by default type-checking a file in the README
-  directory fails.
+* By default the README files are not exported in the
+  `standard-library.agda-lib` file in order to avoid
+  clashing with other people's README files.
 
-* If you wish to type-check a README file, then you will need to change the line:
-  ```
-  include: src
-  ```
-  to
-  ```
-  include: src .
-  ```
-  in the `standard-library.agda-lib` file.
-
-* Please do not include this change in your pull request.
-
+* If you wish to type-check a README file, then you will
+  need to change the present working directory to `doc/`
+  where an appropriate `standard-library-doc.agda-lib`
+  file is present.
 
 Continuous Integration (CI)
 ===========================
