+++ CHANGELOG.md
@@ -1,9 +1,9 @@
-Version 2.0-rc1
-===============
+Version 2.0
+===========
 
-The library has been tested using Agda 2.6.4.
+The library has been tested using Agda 2.6.4 and 2.6.4.1.
 
-NOTE: Version `2.0` contains various breaking changes and is not backwards 
+NOTE: Version `2.0` contains various breaking changes and is not backwards
 compatible with code written with version `1.X` of the library.
 
 Highlights
@@ -33,9 +33,15 @@ Highlights
 Bug-fixes
 ---------
 
+* In `Algebra.Structures` the records `IsRing` and `IsRingWithoutOne` contained an unnecessary field
+  `zero : RightZero 0# *`, which could be derived from the other ring axioms.
+  Consequently this field has been removed from the record, and the record
+  `IsRingWithoutAnnihilatingZero` in `Algebra.Structures.Biased` has been
+  deprecated as it is now identical to is `IsRing`.
+  
 * In `Algebra.Definitions.RawSemiring` the record `Prime` did not
-  enforce that the number was not divisible by `1#`. To fix this 
-  `p∤1 : p ∤ 1#` hsa been added as a field.
+  enforce that the number was not divisible by `1#`. To fix this
+  `p∤1 : p ∤ 1#` has been added as a field.
 
 * In `Data.Container.FreeMonad`, we give a direct definition of `_⋆_` as an inductive
   type rather than encoding it as an instance of `μ`. This ensures Agda notices that
@@ -45,24 +51,24 @@ Bug-fixes
 
 * In `Data.Fin.Properties` the `i` argument to `opposite-suc` was implicit
   but could not be inferred in general. It has been made explicit.
-  
-* In `Data.List.Membership.Setoid` the operations `_∷=_` and `_─_` 
+
+* In `Data.List.Membership.Setoid` the operations `_∷=_` and `_─_`
   had an extraneous `{A : Set a}` parameter. This has been removed.
 
-* In `Data.List.Relation.Ternary.Appending.Setoid` the constructors 
-  were re-exported in their full generality which lead to unsolved meta 
+* In `Data.List.Relation.Ternary.Appending.Setoid` the constructors
+  were re-exported in their full generality which lead to unsolved meta
   variables at their use sites. Now versions of the constructors specialised
   to use the setoid's carrier set are re-exported.
-  
-* In `Data.Nat.DivMod` the parameter `o` in the proof `/-monoˡ-≤` was 
+
+* In `Data.Nat.DivMod` the parameter `o` in the proof `/-monoˡ-≤` was
   implicit but not inferrable. It has been changed to be explicit.
 
-* In `Data.Nat.DivMod` the parameter `m` in the proof `+-distrib-/-∣ʳ` was 
+* In `Data.Nat.DivMod` the parameter `m` in the proof `+-distrib-/-∣ʳ` was
   implicit but not inferrable, while `n` is explicit but inferrable.
   They have been to explicit and implicit respectively.
 
-* In `Data.Nat.GeneralisedArithmetic` the `s` and `z` arguments to the 
-  following functions were implicit but no inferrable:
+* In `Data.Nat.GeneralisedArithmetic` the `s` and `z` arguments to the
+  following functions were implicit but not inferrable:
   `fold-+`, `fold-k`, `fold-*`, `fold-pull`. They have been made explicit.
 
 * In `Data.Rational(.Unnormalised).Properties` the module `≤-Reasoning`
@@ -79,9 +85,9 @@ Bug-fixes
   _≈⟨_⟨_ ↦ _≃⟨_⟨_
   ```
 
-* In `Function.Construct.Composition` the combinators 
+* In `Function.Construct.Composition` the combinators
[…]
