+++ fix-whitespace.yaml
@@ -5,7 +5,6 @@ included-dirs:
 included-files:
   - "*.agda"
   - "*.md"
-  - ".travis.yml"
 
 excluded-files:
   - "README/Text/Tabular.agda"

[…]
