### 2025-11-03 06252d5
Revert "Fast-forward v2.0-release (branch) to v2.0 (tag) (#2856)"
This reverts commit 381b99671bab3ed9f391f41f664ae5f3e4903615.
