+++ README.md
@@ -1,6 +1,6 @@
-[![Ubuntu build](https://github.com/agda/agda-stdlib/actions/workflows/ci-ubuntu.yml/badge.svg)](https://github.com/agda/agda-stdlib/actions/workflows/ci-ubuntu.yml)
+![Travis (.org) branch](https://img.shields.io/travis/agda/agda-stdlib/master?label=master)
+![Travis (.org) branch](https://img.shields.io/travis/agda/agda-stdlib/experimental?label=experimental)
 
-[![Ubuntu build](https://github.com/agda/agda-stdlib/actions/workflows/ci-ubuntu.yml/badge.svg?branch=experimental)](https://github.com/agda/agda-stdlib/actions/workflows/ci-ubuntu.yml)
 
 The Agda standard library
 =========================
@@ -18,19 +18,23 @@ If you're looking to find your way around the library, there are several
 different ways to get started:
 
 - The library's structure and the associated design choices are described
-in the [README.agda](https://github.com/agda/agda-stdlib/tree/master/doc/README.agda).
+in the [README.agda](https://github.com/agda/agda-stdlib/tree/master/README.agda).
 
-- The [README folder](https://github.com/agda/agda-stdlib/tree/master/doc/README),
+- The [README folder](https://github.com/agda/agda-stdlib/tree/master/README),
 which mirrors the structure of the main library, contains examples of how to
 use some of the more common modules. Feel free to [open a new issue](https://github.com/agda/agda-stdlib/issues/new) if there's a particular module you feel could do with
 some more documentation.
 
-- You can [browse the library's source code](https://agda.github.io/agda-stdlib/)
+- You can [browse the library's source code](https://agda.github.io/agda-stdlib/README.html)
 in glorious clickable HTML.
+
+- Finally, you can get an overview of the entire library by looking at the
+[index](https://agda.github.io/agda-stdlib/), which lists all modules
+in the library except the deprecated ones.
 
 ## Installation instructions
 
-See the [installation instructions](https://github.com/agda/agda-stdlib/blob/master/doc/installation-guide.md) for the latest version of the standard library.
+See the [installation instructions](https://github.com/agda/agda-stdlib/blob/master/notes/installation-guide.md) for the latest version of the standard library.
 
 #### Old versions of Agda
 
