+++ dev/README.md
@@ -1,4 +1,0 @@
-# Development playground
-
-This directory allows you to develop modules against the current dev
-version of the stdlib as it currently sits in `src/`.

[…]
