### 2025-11-05 e2ff677
refactor: alternative to #2852 and #2854
