### 2025-11-05 bd5cfee
add: sub-`Bimodule`s
