### 2025-11-05 c05773b
add: `Ideal`s and quotients of `Ring`s
