### 2025-11-05 7238414
update notation based on discussion
