### 2025-11-06 0d16e23
Add various relations over non-empty lists
