+++ CHANGELOG.md
@@ -93,6 +93,14 @@ New modules
 
 * `Effect.Monad.Random` and `Effect.Monad.Random.Instances` for an mtl-style randomness monad constraint.
 
+* Various additions over non-empty lists:
+  ```
+  Data.List.NonEmpty.Relation.Binary.Pointwise
+  Data.List.NonEmpty.Relation.Unary.Any
+  Data.List.NonEmpty.Membership.Propositional
+  Data.List.NonEmpty.Membership.Setoid
+  ```
+
 Additions to existing modules
 -----------------------------
 
@@ -157,6 +165,11 @@ Additions to existing modules
 
   search-least⟨_⟩  : Decidable P → Π[ ∁ P ] ⊎ Least⟨ P ⟩
   search-least⟨¬_⟩ : Decidable P → Π[ P ] ⊎ Least⟨ ∁ P ⟩
+  ```
+
+* In `Data.List.NonEmpty.Relation.Unary.All`:
+  ```
+  map : P ⊆ Q → All P xs → All Q xs
   ```
 
 * In `Data.Nat.ListAction.Properties`
