### 2025-11-06 7ad2b22
add: `Center` of a `Group`
