### 2025-11-08 7a8d3cf
add: functoriality proofs
