### 2025-11-10 c4cea19
Fix changelog typo
