### 2025-11-10 75565ff
Add subgroups and submodules (#2852)
* Add subgroups and submodules

* Let Agda fill in a record for me

* Try different names for the underlying structure

* Bimodule rather than module

* style
