### 2025-11-11 agda/agda-stdlib#2867 @JacquesCarette
System.Directory.Primitive - XdgDirectory problem
This is related to v2.0 and [other's trouble with running the tests](https://github.com/openjournals/joss-reviews/issues/9241).

So the released stdlib 2.0 has
https://github.com/agda/agda-stdlib/blob/2b8fff10f4033b91a6df4007e4a65cb10047c89c/src/System/Directory/Primitive.agda#L29

whereas released stdlib 2.1 has
https://github.com/agda/agda-stdlib/blob/039e1f4abda49f307f7ed1f5e15232a1c8e9c48c/src/System/Directory/Primitive.agda#L29

`git blame` says that this was fixed by @gallais in https://github.com/agda/agda-stdlib/commit/307152fd7a5e294225d68c63784a6af272eb0fc1

This seems to be a GHC synchronization issue (that I have not tracked down in full) where `XdgState` was added. This is new to `directory-1.3.7.0` (which itself is quite old) but it is hard to tell when that became the default for Agda / MAlonzo.

Best guess: v2.0 was released without this particular test suite being run.  Probably because the release instructions do not have a `make testsuite` step! @MatthewDaggitt
