### 2025-11-11 c3d32b1
Update CHANGELOG for the new functions
