### 2025-11-11 agda/agda-stdlib#2867 @gallais
> This seems to be a GHC synchronization issue (that I have not tracked down in full) where XdgState was added. This is new to directory-1.3.7.0 (which itself is quite old) but it is hard to tell when that became the default for Agda / MAlonzo.

Can we force 2.0 to use directory<1.3.7.0? I don't know that there's any easy way to do that because
Agda does not really have a good story when it comes to building projects with dependencies
on Haskell libs: https://github.com/agda/agda/issues/4627

At the momment IIRC we install the libs globally and hope for the best that they're the right
sort of version. :s

I suppose we could hand-cook a .cabal file just for the test suite.
