### 2025-11-11 agda/agda-stdlib#2867 @MatthewDaggitt
So unfortunately we can't backpatch to the v2.0 tag as that will break all sorts of CI scripts that depend on v2.0 of the library, e.g. Wen's github actions. If we back patched it to the v2.0 branch we'd then have to release v2.0.1 which would be a little bit odd this late in the day... I guess we can ask the reviewers of the paper whether they're happy for the test suite to work for current versions but not for the released 2.0?
