### 2025-11-11 6b8b321
Adds "pushforward" to Relation.Unary (#2840)
* add predicate pushforward

* update changelog

* update notation based on discussion

* Additional description

Co-authored-by: jamesmckinna <31931406+jamesmckinna@users.noreply.github.com>

---------

Co-authored-by: jamesmckinna <31931406+jamesmckinna@users.noreply.github.com>
