### 2025-11-12 cbfc19e
add: `Algebra.Definitions.Central`
