### 2025-11-19 4b50bf3
[ add ] `Algebra.Definitions.Central` plus `Consequences` for `Identity` and `Zero` (#2873)
* add: `Algebra.Definitions.Central`

* add: `Identity` and `Zero` elements are `Central`

* Update src/Algebra/Definitions.agda

Co-authored-by: Jacques Carette <carette@mcmaster.ca>

* fix: line break

---------

Co-authored-by: Jacques Carette <carette@mcmaster.ca>
