+++ tests/Makefile
@@ -1,13 +1,17 @@
-INTERACTIVE ?= --interactive
+INTERACTIVE   ?= --interactive
+GHC_EXEC      ?= ghc
+CABAL         ?= cabal
+CABAL_OPTIONS ?= --with-compiler $(GHC_EXEC)
 
 runtests: runtests.agda
 	rm -f _build/runtests
 	rm -f _build/MAlonzo/Code/Qruntests*
-	$(AGDA) --compile-dir=_build/ -c runtests.agda
+	$(AGDA) --compile-dir=_build/ -c --ghc-dont-call-ghc runtests.agda
+	$(CABAL) build $(CABAL_OPTIONS)
 
 test: runtests
-	./_build/runtests $(AGDA_EXEC) $(INTERACTIVE) --timing --failure-file failures --only $(only)
+	$(CABAL) exec $(CABAL_OPTIONS) runtests -- $(AGDA_EXEC) $(INTERACTIVE) --timing --failure-file failures --only $(only)
 
 
 retest: runtests
-	./_build/runtests $(AGDA_EXEC) $(INTERACTIVE) --timing --failure-file failures --only-file failures --only $(only)
+	$(CABAL) exec $(CABAL_OPTIONS) runtests -- $(AGDA_EXEC) $(INTERACTIVE) --timing --failure-file failures --only-file failures --only $(only)
