### 2025-11-19 46d8280
add: `Centre` of an algebra, following #2863
