### 2025-11-19 68f96e5
[ fix ] building the data/appending test
Now that this is sorted, the other ones should be easy
