### 2025-11-19 49dc268
[ ci ] run revamped testsuite
