### 2025-11-19 3cacdde
[ ci ] enable build on v2.0-joss-submission branch
