### 2025-11-19 8dbf96c
[ ci ] haskell/actions/setup -> haskell-actions/setup
