### 2025-11-19 4c746a3
[ refactor ] removing a lot of duplicate code
