### 2025-11-19 7e48fd0
[ ci ] fix GHC call
