### 2025-11-19 b75de2a
[ fix ] regex test case
