### 2025-11-19 00db5c2
add: centre of a `Ring`
