### 2025-11-20 3dc4b35
fix: remove `module ι`, and better description in `CHANGELOG`
