+++ CHANGELOG.md
@@ -106,8 +106,8 @@ New modules
   ```
 
 * `Relation.Binary.Morphism.Construct.On`: given a relation `_∼_` on `B`,
-  and a function `f : A → B`, lift to various `IsRelHomomorphism`s between
-  `_∼_ on f` and `_∼_`.
+  and a function `f : A → B`, construct the canonical `IsRelMonomorphism`
+  between `_∼_ on f` and `_∼_`, witnessed by `f` itself.
 
 Additions to existing modules
 -----------------------------
