### 2025-11-20 8985324
[ ci ] Don't pre-install clock
