+++ .github/workflows/ci-ubuntu.yml
@@ -169,7 +169,6 @@ jobs:
 
       - name: Golden testing
         run: |
-          ${{ env.CABAL_INSTALL }} clock
           make testsuite INTERACTIVE='' AGDA_EXEC='~/.cabal/bin/agda' GHC_EXEC=ghc
 
 
