### 2025-11-20 74114fb
[ doc ] adding a README file
