+++ tests/_config/config.sh
@@ -5,17 +5,12 @@
 # It provides a number of default config options corresponding
 # to the compiler versions the stdlib is being tested with
 #
-# Usage: . PATH/TO/config.sh
+# Usage: . ../../config.sh
 
 set -e
 
-if [ -z ${AGDA_EXEC} ]; then
-    export AGDA_EXEC=agda-2.6.4
-fi
-
-if [ -z ${GHC_EXEC} ]; then
-  export GHC_EXEC=ghc-9.2.8
-fi
+# Ugh, paths are relative to the script sourcing this file!
+. ../../_config/version-numbers.sh
 
 goldenTest () {
 
