### 2025-11-20 cc608e4
[ fix ] add generic support for input files
