### 2025-11-20 #9241 @vidsinghal
> Thanks for the feedback [@vidsinghal](https://github.com/vidsinghal)!
>
> > 1.) The installation instructions are too general, ideally I would like to see a list of dependencies and the exact version requirement for a particular release.
>
> Getting the standard library running depends on nothing but Agda. Although installing Agda of course may require other tools (e.g. cabal), it probably shouldn't be the library's job to tell users how to install the underlying language. Each version's installation instructions (e.g. for v2.0: https://github.com/agda/agda-stdlib/blob/v2.0-joss-submission/doc/installation-guide.md) comes with precise instructions for which versions of Agda you need and this is linked to on the front-page.
>
> Could you provide a little more detail on what you would like to see these instructions include?

Ideally, I would like to have an installation script that installs the project, including the dependencies automatically,
In addition to sourcing the relevant paths etc.

>
> > 2.) For functionality, I would like to see more examples of how to use the library, maybe a small example and trying to prove it using the library would be ideal.
>
> Have you seen the [README](https://github.com/agda/agda-stdlib/tree/master/doc) directory? Within that directory, there's many examples of how to use the most common modules to prove basic concepts in the library:
>
>     * https://github.com/agda/agda-stdlib/blob/master/doc/README/Data/Nat.agda
>
>     * https://github.com/agda/agda-stdlib/blob/master/doc/README/Data/Integer.agda
>
>     * https://github.com/agda/agda-stdlib/blob/master/doc/README/Data/List/Relation/Unary/Any.agda
>
>
> This includes a [subfolder](https://github.com/agda/agda-stdlib/tree/master/doc/README/Design) containing extensive documentation for general design decisions, e.g.:
>
>     * https://github.com/agda/agda-stdlib/blob/master/doc/README/Design/Hierarchies.agda
>
>
> Also because it's the standard library, it is a little hard to disentangle tutorials of how to use the library vs tutorials of how to use Agda. For example, explanations of how to use the standard library for Python tend to just be subsumed into explanations of how to use Python. There's several excellent full length tutorials that go into how to use Agda and therefore the library. For example, the majority of Section 1 of [PLFA](https://plfa.github.io/) goes through each of the definitions in the library and how can they be used to prove things, and then points to the user to where they are found in the library at the end of each chapter. Would linking to these sort of tutorials satisfy you?

For examples, yes, I think linking the above in the paper would be great.
