### 2025-11-20 f8c17d4
[ misc ] various fixes & using a unique cabal file
