+++ tests/_config/config.sh
@@ -15,7 +15,7 @@ set -eu
 goldenTest () {
 
   AGDA=$1
-  TEST_NAME=$2
+  TEST_NAME="stdlib-test-$2"
 
   # Remember whether the script has an input -- ugh
   if [ -f input ]; then
@@ -43,7 +43,7 @@ goldenTest () {
   ln -sf "$CABAL_BUILD_DIR" dist-newstyle
 
   # Compile the Agda module and build the generated code
-  "$AGDA" --library-file=../../_config/libraries --compile-dir=_build -c --ghc-dont-call-ghc Main.agda > logs/agda-build
+  $AGDA --library-file=../../_config/libraries --compile-dir=_build -c --ghc-dont-call-ghc Main.agda > logs/agda-build
   cabal build "$TEST_NAME" --with-compiler "$GHC_EXEC" > logs/cabal-build
 
   # Run the test
