### 2025-11-21 b53aa9c
[ doc ] for various bits
