### 2025-11-21 b8ce24b
[ add ] `Relation.Binary.Morphism.Construct.On` (#2872)
* add: `Relation.Binary.Morphism.Construct.On`

* fix: top-level descriptive comment

* fix: remove `module ι`, and better description in `CHANGELOG`
