### 2025-11-21 320220b
fix: #2882 (#2883)
