### 2025-11-24 bcef47e
[ add ] left- and right- actions induced by a `Magma` homomorphism
