### 2025-11-25 9bc4710
fix: `CHANGELOG`
