+++ CHANGELOG.md
@@ -86,7 +86,7 @@ New modules
 
 * `Algebra.Construct.Centre.X` for the definition of the centre of an algebra,
   where `X = {Semigroup|Monoid|Group|Ring}`, based on an underlying type defined
-  in `Algebra.Construct.Centre`.
+  in `Algebra.Construct.Centre.Center`.
 
 * `Algebra.Construct.Sub.Group` for the definition of subgroups.
 
