### 2025-11-25 273d05a
[ new ] install script for the standard library
