+++ stdlib-install.sh
@@ -1,0 +1,141 @@
+#!/bin/sh
+
+set -eu
+
+throwError() {
+  echo "\033[91m✗\033[0m $1." >&2
+  exit 1
+}
+
+logHappy() {
+  echo "\033[32m✔\033[0m $1"
+}
+
+# Pick the Agda executable to analyse
+# unless the caller has specified one
+if [ -z ${AGDA_EXEC-} ]; then
+    read -p "What's the name of your Agda executable (default: agda)? " AGDA_EXEC
+    if [ -z "$AGDA_EXEC" ]; then
+        AGDA_EXEC=agda
+    fi
+fi
+
+# Double check that the command exists
+if ! [ -x "$(command -v $AGDA_EXEC)" ]; then
+  throwError "'$AGDA_EXEC' is not a valid executable"
+fi
+
+logHappy "Agda executable: $AGDA_EXEC"
+
+# Ask the executable for its version number
+# unless the caller has specified one
+if [ -z ${AGDA_VERSION-} ]; then
+  AGDA_VERSION=$($AGDA_EXEC --version | head -n 1 | sed "s/^[a-zA-Z ]*\(2[0-9.]*\)\(-.*\)*$/\1/")
+fi
+
+# Double check that the version number is correct
+if ! echo "$AGDA_VERSION" | grep -Eq "^2(\.[0-9]+)+$"; then
+  throwError "'$AGDA_VERSION' is not a valid version number"
+fi
+
+logHappy "Agda version number: $AGDA_VERSION"
+
+# Pick the install directory
+# unless the caller has specified one
+if [ -z ${AGDA_DIR-} ]; then
+  AGDA_DIR=$($AGDA_EXEC --print-agda-app-dir | head -n 1 || true)
+  if echo "$AGDA_DIR" | grep -Eq "^Error.*$"; then
+    AGDA_DIR="$HOME/.agda"
+  fi
+  read -p "Where do you want to install the library (default: $AGDA_DIR)? " AGDA_DIR_OVERWRITE
+  if ! [ -z "$AGDA_DIR_OVERWRITE" ]; then
+    AGDA_DIR="$AGDA_DIR_OVERWRITE"
+  fi
+fi
+
+logHappy "Agda directory: $AGDA_DIR"
+
+if [ -z ${STDLIB_VERSION-} ]; then
+  case "$AGDA_VERSION" in
+    "2.8.0")
+      STDLIB_VERSION=2.3
+      ;;
+    "2.7.0.1")
+      STDLIB_VERSION=2.3
+      ;;
+    "2.7.0")
+      STDLIB_VERSION=2.2
+      ;;
+    "2.6.4.3")
+      STDLIB_VERSION=2.1
+      ;;
+    "2.6.4.2")
+      STDLIB_VERSION=2.0
+      ;;
+    "2.6.4.2")
+      STDLIB_VERSION=2.0
+      ;;
+    "2.6.4.1")
+      STDLIB_VERSION=2.0
+      ;;
+    "2.6.4")
+      STDLIB_VERSION=2.0
+      ;;
+    "2.6.3")
+      STDLIB_VERSION=1.7.3
+      ;;
+    "2.6.2.2")
+      STDLIB_VERSION=1.7.1
+      ;;
+    "2.6.2.1")
+      STDLIB_VERSION=1.7.1
+      ;;
+    "2.6.1")
+      STDLIB_VERSION=1.7.1
+      ;;
+    *)
+      STDLIB_VERSION=experimental
+      ;;
+  esac
+fi
+
+logHappy "Standard library version: $STDLIB_VERSION"
+
+case "$STDLIB_VERSION" in
+  "master")
[…]
