### 2025-11-26 a431f77
[ fix ] add test runner for v2.0-joss-submission (#2884)
