### 2025-11-26 659cc9c
[ fix ] add test runner (#2884)
