### 2025-11-27 0c42bd9
Address @vidsinghal's comments
