+++ paper/paper.bib
@@ -328,5 +328,11 @@ numpages = {13},
 keywords = {type classes, instance arguments, agda, ad hoc polymorphism}
 }
 
-  
-
+@inproceedings{wadler2018programming,
+  title={Programming language foundations in agda},
+  author={Wadler, Philip},
+  booktitle={Brazilian Symposium on Formal Methods},
+  pages={56--73},
+  year={2018},
+  organization={Springer}
+}
