### 2025-12-08 4dce66b
Add map-removeAt to Data.Vec.Properties
