### 2025-12-09 d8a2137
Fix whitespace in changelog
