### 2025-12-09 49567fe
Add map-removeAt to Data.Vec.Properties (#2899)
* Add map-removeAt to Data.Vec.Properties

* Fix whitespace in changelog
