### 2025-12-10 4478a5d
[ test ] update bounds
This is probably not going to build (missing bytestring / random
dependencies in the sample cabal file).
