### 2025-12-10 5be3e36
[ cleanup ]
* CI: remove JOSS build
* Test runner: print test name without buffering
* Script: fix issue leading to nested _build/_build/...
