+++ .github/workflows/ci-ubuntu.yml
@@ -9,7 +9,6 @@ on:
     branches:
       - master
       - experimental
-      - v2.0-joss-submission
   merge_group:
 
 ########################################################################
