### 2025-12-10 827c2fe
[ fix ] touch files before grepping them
