### 2025-12-10 5e9fe26
[ doc ] explain option choice
