### 2025-12-10 5866910
[ more ] check dependencies + debug mode
