### 2025-12-10 a45495d
[ fix ] be more careful about overwriting dirs
