### 2025-12-10 eba7fa7
[ ergonomics ] make more of an effort
Keep asking the user until we get a valid Agda executable
