### 2025-12-10 b7e5775
[ doc ] the installation instruction
