### 2025-12-10 7c4e75f
[ cleanup ] follow the shellcheck advice
