+++ stdlib-install.sh
@@ -6,35 +6,27 @@
 set -eu
 
 isDebugMode() {
-  ! [ -z ${DEBUG-} ]
+  [ -n "${DEBUG-}" ]
 }
 
-if ! [ -z ${SHELL_DEBUG-} ]; then
+if [ -n "${SHELL_DEBUG-}" ]; then
   set -x
 fi
 
 throwError() {
-  echo "\033[91m✗\033[0m $1." >&2
+  printf "\033[91m✗\033[0m %s.\n" "$1" >&2
   exit 1
 }
 
-logHappy() {
-  echo "\033[32m✔\033[0m $1"
-}
-
-logUnhappy() {
-  echo "\033[91m✗\033[0m $1."
-}
-
-logWarning() {
-  echo "\033[93m⚠\033[0m $1"
-}
+logHappy()   { printf "\033[32m✔\033[0m %s.\n" "$1"; }
+logUnhappy() { printf "\033[91m✗\033[0m %s.\n" "$1"; }
+logWarning() { printf "\033[93m⚠\033[0m %s.\n" "$1"; }
 
 ########################################################################
 ## PRELIMINARY CHECKS
 
 checkDependency () {
-  if ! [ -x "$(command -v $1)" ]; then
+  if ! [ -x "$(command -v "$1")" ]; then
     throwError "Missing dependency: I could not find the executable '$1'"
   elif isDebugMode; then
     logHappy "Found '$1'"
@@ -54,20 +46,21 @@ checkDependency "wget"
 
 # Pick the Agda executable to analyse
 # unless the caller has specified one
-if [ -z ${AGDA_EXEC-} ]; then
+if [ -z "${AGDA_EXEC-}" ]; then
   while true; do
-    read -p "What's the name of your Agda executable (default: agda)? " AGDA_EXEC
+    printf "What's the name of your Agda executable (default: agda)? "
+    read -r AGDA_EXEC
     if [ -z "$AGDA_EXEC" ]; then
       AGDA_EXEC=agda
     fi
     # Double check that the command exists
-    if ! [ -x "$(command -v $AGDA_EXEC)" ]; then
+    if ! [ -x "$(command -v "$AGDA_EXEC")" ]; then
       logUnhappy "'$AGDA_EXEC' is not a valid executable"
     else
       break
     fi
   done
-elif ! [ -x "$(command -v $AGDA_EXEC)" ]; then
+elif ! [ -x "$(command -v "$AGDA_EXEC")" ]; then
   throwError "'$AGDA_EXEC' is not a valid executable"
 fi
 
@@ -75,7 +68,7 @@ logHappy "Agda executable: $AGDA_EXEC"
 
 # Ask the executable for its version number
 # unless the caller has specified one
-if [ -z ${AGDA_VERSION-} ]; then
+if [ -z "${AGDA_VERSION-}" ]; then
   # There is a more recent "--numeric-version" option but old
   # versions of Agda do not implement it!
   AGDA_VERSION=$($AGDA_EXEC --version | head -n 1 | sed "s/^[a-zA-Z ]*\(2[0-9.]*\)\(-.*\)*$/\1/")
@@ -90,26 +83,26 @@ logHappy "Agda version number: $AGDA_VERSION"
 
 # Pick the install directory
 # unless the caller has specified one
-if [ -z ${AGDA_DIR-} ]; then
+if [ -z "${AGDA_DIR-}" ]; then
   AGDA_DIR=$($AGDA_EXEC --print-agda-app-dir | head -n 1 || true)
[…]
