### 2025-12-10 1f03bf6
[ admin ] add a step to the release guide
