### 2025-12-11 1b6ff01
[ more ] gracefully report that wget failed
