### 2025-12-11 6e055c0
[ cleanup ] wording
