### 2025-12-11 d0eca1f
[ cleanup ] use variable rather than repetition
