### 2025-12-11 4a762a6
[ doc ] more structure in the README
