+++ README.md
@@ -30,12 +30,16 @@ in glorious clickable HTML.
 
 ## Installation instructions
 
+### Automated installation (currently experimental)
+
 On Unix-style systems, it should be as simple as running the following command in
 a shell:
 
 ```shell
 sh -c "$(curl --proto '=https' --tlsv1.2 -s https://raw.githubusercontent.com/agda/agda-stdlib/refs/heads/master/stdlib-install.sh))"
 ```
+
+### Manual installation
 
 See the [installation instructions](https://github.com/agda/agda-stdlib/blob/master/doc/installation-guide.md) for the latest version of the standard library.
 
