### 2025-12-12 732160f
[ fix ] remove redundant full stop
