### 2025-12-12 664159e
[ fix ] defaults-X.Y.Z is not supported yet
I'll open a PR to Agda to add it to the next releases.
