### 2025-12-15 #9241 @danielskatz
@dorchard - I see that you still have the "automated tests" item unchecked. Can you check this off now? If not, what do you think is needed?
