### 2025-12-15 #9241 @dorchard
I'm still having problems - I have posted the issue here: https://github.com/agda/agda-stdlib/pull/2896#issuecomment-3657172311 Not sure what I am doing wrong.
