### 2025-12-17 960f4c0
minor changes in paper.md
