### 2025-12-19 #9241 @danielskatz
@MatthewDaggitt - please check the proof that this will generate.
