### 2025-12-19 #9241 @danielskatz
@MatthewDaggitt - In your zenodo archive, you might change the other contributors to be authors to match the paper. You also might change the license to be the software repo's license. Note that you can edit the metadata without making a new archive.
