### 2025-12-19 #9241 @MatthewDaggitt
The proof looks good to me. I have updated the Zenodo meta-data as you suggest above (changed contributors to authors, and added the license).
