### 2025-12-19 #9241 @JacquesCarette
I have merged it.
